modification du type de la fonction sous et pareillement pour le type des paramètres de la fonction

+++ sous.java
@@ -1,4 +1,4 @@
 // Soustrait deux nombres entier
-public int sous(int a, int b){
+public Double sous(Double a, Double b){
 	return a-b;
 }
